requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Pillow==10.0.0
opencv-python-headless==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Flask-Mail==0.9.1